requirements.txt
discord.py==2.3.2
pandas==2.2.1
pytz==2023.3.post1
datetime